+++ tests/testthat/test-utils.R
@@ -53,3 +53,10 @@ test_that("print", {
     M = biproporz(matrix(c(51,60,63,98,100,102,45,120,144), nrow = 3), 4:6)
     expect_equal(capture.output(print(M)), capture.output(print(as.matrix(M))))
 })
+
+test_that("collapse_names", {
+    x = c("Abc", "XYZ")
+    y = c(2, 6, 67)
+    expect_equal(collapse_names(x), "'Abc', 'XYZ'")
+    expect_equal(collapse_names(y), "2, 6, 67")
+})